added mobile styles to all the pages
