+++ clinician_overview/static/clinician_overview/client_response_page.css
@@ -1,0 +1,5 @@
+#client-responses-table .etable-cell {
+    white-space: wrap;
+    word-break: break-word;
+    text-align: left;
+}
