+++ clinician_overview/static/clinician_overview/clients_page.css
@@ -6,8 +6,15 @@
   width: 30%;
 }
 
+#clients-page-grid-container .etable .etable-cell-pill.client-tag {
+  white-space: nowrap;
+}
+
 .grey-pill {
     background-color: rgba(224, 224, 224, 1);
+    box-decoration-break: clone;
+    -webkit-box-decoration-break: clone;
+    line-height: 2rem;
 }
 
 .add-client-button {
@@ -85,5 +92,41 @@ input[type="submit"] {
     width: fit-content;
 }
 
+.etable th, .etable td {
+  white-space: wrap !important;
+}
 
+div:has(>#register-client-id-form) {
+  margin: 0 auto;
+}
 
+.etable {
+  min-width: 0;
+}
+
+#selectable-table {
+  overflow-x: auto;
+}
+
+#clients-page-grid-container {
+  display:grid; 
+  margin: 10px auto;
+}
+
+@media (max-width: 768px) { 
+  #clients-page-grid-container {
+    grid-template-columns: 1fr;
+  }
+  div:has(> #register-client-id-form) {
+    margin: 0 auto;
+  }
+}
+
+@media (min-width: 769px) {
+  #clients-page-grid-container {
+    grid-template-columns: 4fr 1fr;
+  }
+  div:has(> #register-client-id-form) {
+    margin-right: 2em;
+  }
+}
