+++ clinician_overview/static/clinician_overview/clinician_base_styles.css
@@ -32,6 +32,51 @@ h1 {
 
 .content-columns {
   display: flex;
+  gap: 2rem;
+}
+
+#client-header-container {
+    background-color: rgba(255, 255, 255, 0.2);
+    padding: 1rem;
+    color: white;
+}
+
+@media (max-width: 768px) {
+  .content-columns {
+    flex-direction: column;
+  }
+
+  #client-header-container {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+  }
+
+  nav ul {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    margin: auto;
+  }
+
+  #main {
+    padding-left: 0;
+    margin-left: 0;
+  }
+
+  .etable-container {
+    overflow-x: auto;
+  }
+}
+@media (min-width: 769px) {
+  .content-columns {
+    flex-direction: row;
+  }
+
+  #main {
+    margin-left: 1rem; 
+    width: 100%;
+  }
 }
 
 a {
@@ -101,11 +146,10 @@ header, .tab-header {
   padding: 1rem;
   border: 2px solid var(--bluish);
   text-align: center;
-  white-space: nowrap;
+  /* white-space: nowrap; */
 }
 
 .etable-container {
-  max-width: 90%;
   margin: auto;
 }
 
@@ -140,6 +184,9 @@ footer {
 .bluish-pill {
     border-radius: 50px;
     background-color: color-mix(in srgb, var(--bluish) 60%, transparent);
+    line-height: 2rem;
+    box-decoration-break: clone;
+    -webkit-box-decoration-break: clone;
 }
 
 select {
@@ -151,13 +198,13 @@ select {
     height: 40px;
     color: rgb(75, 53, 28);
     line-height: 1.4em;
+    max-width: 100%;
 }
 
 .centered-flex-container {
     display: flex;
     align-items: center;
 }
-
 
 .flex-space-between {
     justify-content: space-between;
@@ -171,16 +218,13 @@ select {
     gap: 2em;
 }
 
-
 .hide-when-no-rows-selected {
     display: none;
 }
 
-
 #selected-counter {
     margin-right: 5px;
 }
-
 
 .material-symbols-outlined.material-tags-control-badge {
     display: inline;
