+++ clinician_overview/static/clinician_overview/notifications.css
@@ -1,0 +1,4 @@
+.etable-container {
+    max-width: 90%;
+    margin: 0 auto;
+}
