+++ clinician_overview/templates/clinician_overview/client_response_page.html
@@ -1,5 +1,9 @@
 {% extends 'clinician_overview/client_base_header_and_tags_layout.html' %}
 {% load static %}
+{% block head %}
+<link rel="stylesheet" href="{% static 'clinician_overview/client_response_page.css' %}">
+<link rel="stylesheet" href="{% static 'clinician_overview/client_summary_page.css' %}">
+{% endblock %}
 {% block title %} Client Response {% endblock %}
 
 {% block client-data %}
@@ -9,24 +13,26 @@
 <span>
   {{html_result_display}}
 </span>
-<table class="etable">
-  <thead>
-    <tr class="etable-header-row">
-      <th class="etable-header-cell">Question</th>
-      <th class="etable-header-cell">Response</th>
+<div class="etable-container">
+  <table class="etable" id="client-responses-table">
+    <thead>
+      <tr class="etable-header-row">
+        <th class="etable-header-cell">Question</th>
+        <th class="etable-header-cell">Response</th>
+      </tr>
+    </thead>
+    {% for item in response_items %}
+    <tr class="etable-row">
+      <td class="etable-cell">{{item.question.text}}</td>
+      <td class="etable-cell">
+        {% if item.answerID %}
+          {{item.answerID.text}}
+        {% else %}
+          {{item.answer}}
+        {% endif %}
+      </td>
     </tr>
-  </thead>
-  {% for item in response_items %}
-  <tr class="etable-row">
-    <td class="etable-cell">{{item.question.text}}</td>
-    <td class="etable-cell">
-      {% if item.answerID %}
-        {{item.answerID.text}}
-      {% else %}
-        {{item.answer}}
-      {% endif %}
-    </td>
-  </tr>
-  {% endfor %}
-</table>
+    {% endfor %}
+  </table>
+</div>
 {% endblock %}
