+++ clinician_overview/templates/clinician_overview/clinician_base_layout.html
@@ -38,7 +38,7 @@
             <li class="{% if nav_section == 'help' %}active{% endif %}"><a>Help</a></li>
           </ul>
         </nav>
-        <div style="margin-left: 1rem; width: 100%">
+        <div id="main">
           {% block content %}
           <p>placeholder</p>
           {% endblock %}
