+++ clinician_overview/templates/clinician_overview/detailed_form_result_display/dest_results.html
@@ -41,4 +41,3 @@
         Indicate pronounced or significant dissociative experiences. These scores may warrant a more comprehensive assessment to determine the most appropriate intervention or therapeutic approach.
     </li>
 </ul>
-</div>
