+++ clinician_overview/templates/clinician_overview/notifications_page.html
@@ -1,5 +1,7 @@
 {% extends 'clinician_overview/clinician_base_layout.html' %} {% load static %}
-{% block title %} Notifications {% endblock %} {% block head %}
+{% block title %} Notifications {% endblock %} 
+{% block head %}
+<link rel="stylesheet" href="{% static 'clinician_overview/notifications.css' %}">
 {% endblock %} {% block content %}
 <div class="page-white-background">
   <header>
